[Lights] Fixed a bug where transparent shadows would not always work

+++ Runtime/Audio/Audio.cpp
@@ -131,7 +131,7 @@ namespace Spartan
         return true;
     }
 
-    void Audio::OnTick(float delta_time)
+    void Audio::OnTick(double delta_time)
     {
         // Don't play audio if the engine is not in game mode
         if (!m_context->m_engine->EngineMode_IsSet(Engine_Game))
