+++ Runtime/World/Components/Camera.h
@@ -55,7 +55,7 @@ namespace Spartan
 
         //= ICOMPONENT ===============================
         void OnInitialize() override;
-        void OnTick(float delta_time) override;
+        void OnTick(double delta_time) override;
         void Serialize(FileStream* stream) override;
         void Deserialize(FileStream* stream) override;
         //============================================
@@ -129,7 +129,7 @@ namespace Spartan
         Math::Matrix ComputeProjection(const bool reverse_z, const float near_plane = 0.0f, const float far_plane = 0.0f);
 
     private:
-        void FpsControl(float delta_time);
+        void FpsControl(double delta_time);
 
         float m_aperture                    = 50.0f;        // Size of the lens diaphragm (mm). Controls depth of field and chromatic aberration.
         float m_shutter_speed               = 1.0f / 60.0f; // Length of time for which the camera shutter is open (sec). Also controls the amount of motion blur.
