+++ Runtime/World/Components/Light.h
@@ -69,7 +69,7 @@ namespace Spartan
         //= COMPONENT ================================
         void OnInitialize() override;
         void OnStart() override;
-        void OnTick(float delta_time) override;
+        void OnTick(double delta_time) override;
         void Serialize(FileStream* stream) override;
         void Deserialize(FileStream* stream) override;
         //============================================
