+++ Runtime/World/World.h
@@ -44,7 +44,7 @@ namespace Spartan
 
         //= ISubsystem ======================
         bool OnInitialise() override;
-        void OnTick(float delta_time) override;
+        void OnTick(double delta_time) override;
         //===================================
         
         void New();
